+++ src/LASRstages/addrgb.h
@@ -11,4 +11,13 @@ public:
   LASRaddrgb* clone() const override { return new LASRaddrgb(*this); };
 };
 
+class LASRremovergb: public Stage
+{
+public:
+  bool process(PointCloud*& las) override { return las->remove_rgb(); };
+  std::string get_name() const override { return "remove_rgb"; };
+  LASRremovergb* clone() const override { return new LASRremovergb(*this); };
+};
+
+
 #endif