+++ src/LASRstages/writelas.cpp
@@ -25,6 +25,21 @@ bool LASRlaswriter::set_parameters(const nlohmann::json& stage)
   keep_buffer = stage.value("keep_buffer", false);
   copc_density = stage.value("density", 256);
   copc_depth = stage.value("max_depth", -1);
+  version_minor = stage.value("version", 0xFF); // 0xFF auto-detect
+  point_format = stage.value("pdrf", 0xFF);
+
+  if (version_minor != 0xFF && version_minor > 4)
+  {
+    last_error = "Invalid version for LAS format";
+    return false;
+  }
+
+  if (point_format != 0xFF && point_format > 10)
+  {
+    last_error = "Invalid point data format for LAS format";
+    return false;
+  }
+
   return true;
 }
 
@@ -134,7 +149,7 @@ bool LASRlaswriter::process(PointCloud*& las)
   return true;
 }
 
-void LASRlaswriter::set_header(Header*& header)
+bool LASRlaswriter::set_header(Header*& header)
 {
   // We are receiving a new header because a reader start reading a new file
 
@@ -143,13 +158,29 @@ void LASRlaswriter::set_header(Header*& header)
   if (lasio)
   {
     lasio->reset_accessor();
-    return;
+    return true;
   }
 
-  lasio = new LASio();
-  lasio->init(header);
-  lasio->set_copc_max_depth(copc_depth);
-  lasio->set_copc_density(copc_density);
+  // Use a tmp copy of the header to force some option without modifying the original header
+  Header h = *header;
+  h.signature = "LASF";
+  h.point_data_format = point_format;
+  h.version_minor = version_minor;
+
+  try
+  {
+    lasio = new LASio();
+    lasio->init(&h);
+    lasio->set_copc_max_depth(copc_depth);
+    lasio->set_copc_density(copc_density);
+  }
+  catch (const std::exception& e)
+  {
+    last_error = e.what();
+    return false;
+  }
+
+  return true;
 }
 
 bool LASRlaswriter::set_chunk(Chunk& chunk)